Move each expr to its own module
